added user authorization

+++ app/db/todo-postgres.go
@@ -1,7 +1,9 @@
 package db
 
 import (
+	"crypto/md5"
 	"database/sql"
+	"encoding/hex"
 	"fmt"
 
 	_ "github.com/lib/pq"
@@ -56,9 +58,9 @@ func (db *postgresDB) Migrate() error {
 	return nil
 }
 
-func (db *postgresDB) Update(input *models.Todo) (*models.Todo, error) {
-	updateStmt := "UPDATE todos SET task=$1, completed=$2 WHERE id=$3 RETURNING id;"
-	row := db.sql.QueryRow(updateStmt, input.Text, input.Completed, input.Id)
+func (db *postgresDB) Update(input *models.Todo, userId int) (*models.Todo, error) {
+	updateStmt := "UPDATE todos SET task=$1, completed=$2 WHERE id=$3 AND userid=$4 RETURNING id;"
+	row := db.sql.QueryRow(updateStmt, input.Text, input.Completed, input.Id, userId)
 	var id int
 	if err := row.Scan(&id); err != nil {
 		return nil, err
@@ -67,9 +69,10 @@ func (db *postgresDB) Update(input *models.Todo) (*models.Todo, error) {
 	return input, nil
 }
 
-func (db *postgresDB) List(start int, count int) (*models.TodoList, error) {
+func (db *postgresDB) List(start int, count int, userId int) (*models.TodoList, error) {
 	stm := `
 		SELECT id, task, completed FROM todos 
+		WHERE userid=$3
 		ORDER BY id ASC 
 		LIMIT $1 
 		OFFSET $2;
@@ -78,7 +81,7 @@ func (db *postgresDB) List(start int, count int) (*models.TodoList, error) {
 	if err != nil {
 		return nil, err
 	}
-	rows, err := stmt.Query(count, start)
+	rows, err := stmt.Query(count, start, userId)
 	if err != nil {
 		return nil, err
 	}
@@ -104,13 +107,13 @@ func (db *postgresDB) List(start int, count int) (*models.TodoList, error) {
 		list = append(list, todo)
 	}
 
-	row := db.sql.QueryRow("SELECT COUNT(*) FROM todos;")
+	row := db.sql.QueryRow("SELECT COUNT(*) FROM todos WHERE userid=$1;", userId)
 	var totalCount int
 	if err := row.Scan(&totalCount); err != nil {
 		return nil, err
 	}
 
-	row = db.sql.QueryRow("SELECT COUNT(*) FROM todos WHERE completed = true;")
+	row = db.sql.QueryRow("SELECT COUNT(*) FROM todos WHERE completed = true AND userid=$1;", userId)
 	var completedCount int
 	if err := row.Scan(&completedCount); err != nil {
 		return nil, err
@@ -123,8 +126,8 @@ func (db *postgresDB) List(start int, count int) (*models.TodoList, error) {
 	}, nil
 }
 
-func (db *postgresDB) Add(input *models.AddTodoRequest) (*models.Todo, error) {
-	stmnt, err := db.sql.Prepare("INSERT INTO todos(task, completed) VALUES ($1, $2) RETURNING id;")
+func (db *postgresDB) Add(input *models.AddTodoRequest, userId int) (*models.Todo, error) {
+	stmnt, err := db.sql.Prepare("INSERT INTO todos(task, completed, userid) VALUES ($1, $2, $3) RETURNING id;")
 	if err != nil {
 		return nil, err
 	}
@@ -133,7 +136,7 @@ func (db *postgresDB) Add(input *models.AddTodoRequest) (*models.Todo, error) {
 		AddTodoRequest: *input,
 	}
 
-	row := stmnt.QueryRow(input.Text, input.Completed)
+	row := stmnt.QueryRow(input.Text, input.Completed, userId)
 	if err := row.Scan(&res.Id); err != nil {
 		return nil, err
 	}
@@ -141,12 +144,40 @@ func (db *postgresDB) Add(input *models.AddTodoRequest) (*models.Todo, error) {
 	return res, nil
 }
 
-func (db *postgresDB) Delete(id int) error {
-	updateStmt := "DELETE FROM todos WHERE id=$1;"
-	_, err := db.sql.Exec(updateStmt, id)
+func (db *postgresDB) Delete(id int, userId int) error {
+	updateStmt := "DELETE FROM todos WHERE id=$1 AND userid=$2;"
+	_, err := db.sql.Exec(updateStmt, id, userId)
 	if err != nil {
 		return err
 	}
 
 	return nil
+}
+
+func (db *postgresDB) Register(input *models.RegisterRequest) error {
+	query := "INSERT INTO users(login, passwordhash) values($1, $2) RETURNING id;"
+	row := db.sql.QueryRow(query, input.Login, db.hash(*input.Password))
+	var id int
+	if err := row.Scan(&id); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (db *postgresDB) Login(input *models.LoginRequest) (*int, error) {
+	query := "SELECT id FROM users WHERE login=$1 AND passwordhash=$2;"
+	row := db.sql.QueryRow(query, input.Login, db.hash(*input.Password))
+	var id int
+	if err := row.Scan(&id); err != nil {
+		return nil, err
+	}
+
+	return &id, nil
+}
+
+func (db *postgresDB) hash(in string) string {
+	input := fmt.Sprintf("salt_%s_salt", in)
+	res := md5.Sum([]byte(input))
+	return hex.EncodeToString(res[:])
 }
